+++ R/write_lines.R
@@ -17,3 +17,28 @@ write_lines <- function(x, path, na = "NA", append = FALSE) {
 
   invisible(x)
 }
+
+
+#' @rdname write_lines
+#' @export
+write_file <- function(x, path, append = FALSE) {
+  stopifnot(is.character(x))
+
+  path <- normalizePath(path, mustWork = FALSE)
+
+  write_file_(x, path, append = append)
+
+  invisible(x)
+}
+
+#' @rdname write_lines
+#' @export
+write_file_raw <- function(x, path, append = FALSE) {
+  stopifnot(is.raw(x))
+
+  path <- normalizePath(path, mustWork = FALSE)
+
+  write_file_raw_(x, path, append = append)
+
+  invisible(x)
+}