+++ src/write.cpp
@@ -21,3 +21,29 @@ void write_lines_(const CharacterVector &lines, const std::string &path, const s
 
   return;
 }
+
+// [[Rcpp::export]]
+void write_file_raw_(RawVector x, const std::string &path, bool append = false) {
+  std::ofstream output(path.c_str(), append ? std::ofstream::app : std::ofstream::trunc);
+
+  if (output.fail()) {
+    stop("Failed to open '%s'.", path);
+  }
+
+  std::copy(x.begin(), x.end(), std::ostream_iterator<char>(output));
+
+  return;
+}
+
+// [[Rcpp::export]]
+void write_file_(std::string x, const std::string &path, bool append = false) {
+  std::ofstream output(path.c_str(), append ? std::ofstream::app : std::ofstream::trunc);
+
+  if (output.fail()) {
+    stop("Failed to open '%s'.", path);
+  }
+
+  std::copy(x.begin(), x.end(), std::ostream_iterator<char>(output));
+
+  return;
+}